+++ public/app/features/dashboard/components/SaveDashboard/useDashboardSave.tsx
@@ -32,7 +32,7 @@ export const useDashboardSave = (dashboard: DashboardModel, isCopy = false) => {
     async (clone: DashboardModel, options: SaveDashboardOptions, dashboard: DashboardModel) => {
       try {
         const result = await saveDashboard(clone, options, dashboard);
-        dashboard.version = result.version;
+        dashboard.version = result.data.version;
         dashboard.clearUnsavedChanges();
 
         // important that these happen before location redirect below
